Add RuneLite image docker build file.
This Dockerfile gives the recipe for generating an Ubuntu RuneLite
docker image. When launched, the container will automatically spawn the
RuneLite client. A osrs shell script is included in this changeset. The
osrs script is executed by the container instance and runs the RuneLite
JAR file which ultimately spawns the client.